+++ out/functions/createFile.js
@@ -3,8 +3,8 @@ Object.defineProperty(exports, "__esModule", { value: true });
 const vscode = require("vscode");
 const variables_1 = require("../variables");
 async function createFile(name, ext) {
-    let fileUri = vscode.Uri.file(variables_1.default().wsPath + "/" + name + "/" + ext);
-    let file = await vscode.workspace.fs.writeFile(fileUri, variables_1.default().nullText);
+    let fileUri = vscode.Uri.file(variables_1.default().wsPath + "/" + name + ext);
+    await vscode.workspace.fs.writeFile(fileUri, variables_1.default().nullText);
     vscode.workspace.applyEdit(variables_1.default().wsEdit);
     vscode.window.showTextDocument(fileUri);
 }